Move benchmark suites into python dir.
